types: make neurons package strict-clean and remove it from the mypy ledger
Fixes 26 strict errors: type-arg (np.ndarray -> np.ndarray[Any, Any] across 8
biophysical models; tuple[float, ...] for the glif/mihalas rust-arg bundles;
list[int]/list[float] and dict[str, float] for arcane_neuron state) and
no-any-return (float() wraps on numpy scalar rate/gate expressions in
siegert/marder/mainen/hodgkin_huxley; typed ndarray local in the siegert erf
approximation; typed type/dict/str locals in models.__getattr__ and
universal_dsl). Removes one unused type-ignore in _units. isort cleanup.

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)

+++ src/sc_neurocore/neurons/models/marder_stg.py
@@ -9,6 +9,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
+
 import numpy as np
 
 
@@ -52,7 +53,7 @@ class MarderSTGNeuron:
     v_threshold: float = -20.0
 
     def _boltz(self, v: float, v_half: float, k: float) -> float:
-        return 1.0 / (1.0 + np.exp((v_half - v) / k))
+        return float(1.0 / (1.0 + np.exp((v_half - v) / k)))
 
     def step(self, current: float) -> int:
         v_prev = self.v
